refactor: implement barrel export
