+++ app/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './toast-type'
 export * from './tips-event-type'
+export * from './types'
